[IMP] base_rest: Move code to support datamodel into a dedicated addon
