+++ base_rest_datamodel/README.rst
@@ -1,0 +1,115 @@
+===================
+Base Rest Datamodel
+===================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Frest--framework-lightgray.png?logo=github
+    :target: https://github.com/OCA/rest-framework/tree/12.0/base_rest_datamodel
+    :alt: OCA/rest-framework
+.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
+    :target: https://translation.odoo-community.org/projects/rest-framework-12-0/rest-framework-12-0-base_rest_datamodel
+    :alt: Translate me on Weblate
+.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
+    :target: https://runbot.odoo-community.org/runbot/271/12.0
+    :alt: Try me on Runbot
+
+|badge1| |badge2| |badge3| |badge4| |badge5| 
+
+This addon allows you to use DataModel objects as params and/or response with your
+REST API methods.
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Usage
+=====
+
+To use DataModel instances as request and/or response of a REST service endpoint
+you must:
+
+* Define your DataModel objects;
+* Provides the information required to the ``odoo.addons.base_rest.restapi.method`` decorator;
+
+
+.. code-block:: python
+
+    from marshmallow import fields
+
+    from odoo.addons.base_rest import restapi
+    from odoo.addons.component.core import Component
+    from odoo.addons.datamodel.core import Datamodel
+
+    class PingMessage(Datamodel):
+        _name = "ping.message"
+
+        message = fields.String(required=True, allow_none=False)
+
+
+    class PingService(Component):
+        _inherit = 'base.rest.service'
+        _name = 'ping.service'
+        _usage = 'ping'
+        _collection = 'my_module.services'
+
+
+        @restapi.method(
+            [(["/pong"], "GET")],
+            input_param=restapi.Datamodel("ping.message"),
+            output_param=restapi.Datamodel("ping.message"),
+            auth="public",
+        )
+        def pong(self, ping_message):
+            PingMessage = self.env.datamodels["ping.message"]
+            return PingMessage(message = "Received: " + ping_message.message)
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/rest-framework/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/OCA/rest-framework/issues/new?body=module:%20base_rest_datamodel%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* ACSONE SA/NV
+
+Contributors
+~~~~~~~~~~~~
+
+* Laurent Mignon <laurent.mignon@acsone.eu>
+
+Maintainers
+~~~~~~~~~~~
+
+This module is maintained by the OCA.
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: https://odoo-community.org
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+This module is part of the `OCA/rest-framework <https://github.com/OCA/rest-framework/tree/12.0/base_rest_datamodel>`_ project on GitHub.
+
+You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
