+++ base_rest_datamodel/__manifest__.py
@@ -1,0 +1,16 @@
+# Copyright 2020 ACSONE SA/NV
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+{
+    "name": "Base Rest Datamodel",
+    "summary": """
+        Datamodel binding for base_rest""",
+    "version": "12.0.1.0.0",
+    "license": "AGPL-3",
+    "author": "ACSONE SA/NV,Odoo Community Association (OCA)",
+    "website": "https://acsone.eu/",
+    "depends": ["base_rest", "datamodel"],
+    "data": [],
+    "demo": [],
+    "external_dependencies": {"python": ["apispec"]},
+}
