+++ base_rest_datamodel/readme/CONTRIBUTORS.rst
@@ -1,0 +1,1 @@
+* Laurent Mignon <laurent.mignon@acsone.eu>
