+++ base_rest_datamodel/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This addon allows you to use DataModel objects as params and/or response with your
+REST API methods.
